Update README.md
Abstract: This paper describes the design and implementation of a compact FPGA-based controlled ultrasonic transducer system. We aimed to achieve ultrasound beam steering using three elementary ultrasonic transducers. The transducers formed a steerable beam through ultrasound interference, which was controlled by adjusting the phase differences between them. We intended to demonstrate the validation of the system with a graph illustrating a radar-like beamforming. Using elementary transmitters with an original aperture angle of 50°-130°, we achieved an aperture angle of 74°-101°. Finally, we gave an insight into further development possibilities, which could be used to realize a finished device. The soul of the device used was provided by a ZCU102 development panel, which had a big impact on the further development phase later on. The solution offered a compact and inexpensive solution to Non-Destructive Testing (NDT) systems; as a demonstration, we showed its applicability in a wood inspection example.